syncope/pyproject.toml
[tool.poetry]
name = "syncope"
version = "0.1.0"
description = ""
authors = ["Grayson <grayson.miller124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyzmq = "^25.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
syncope = "syncope.__main__:main"
